+++ python/GafferImageTest/ImageAlgoTest.py
@@ -275,52 +275,52 @@ class ImageAlgoTest( GafferImageTest.ImageTestCase ) :
 			numTilesX * numTilesY * 4
 		)
 
-	def testSortChannelNames( self ):
+	def testSortedChannelNames( self ):
 
 		# Sort RGBA
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( [ "R", "A", "B", "G" ] ), [ "R", "G", "B", "A" ] )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( [ "R", "A", "B", "G" ] ), [ "R", "G", "B", "A" ] )
 
 		# Sort RGBA	before other channels
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( [ "A", "Arc", "Z", "ZSomethingElse", "H", "Bark", "G", "ZBack", "custom", "F", "R", "B" ] ), [ "R", "G", "B", "A", "Arc", "Bark", "F", "H", "Z", "ZBack", "ZSomethingElse", "custom" ] )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( [ "A", "Arc", "Z", "ZSomethingElse", "H", "Bark", "G", "ZBack", "custom", "F", "R", "B" ] ), [ "R", "G", "B", "A", "Arc", "Bark", "F", "H", "Z", "ZBack", "ZSomethingElse", "custom" ] )
 
 		# Sort default layer before named layers
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( [ "A", "G", "A.G", "B.B", "A.R", "B.R", "B", "R", "A.B", "B.G" ] ), [ "R", "G", "B", "A", "A.R", "A.G", "A.B", "B.R", "B.G", "B.B" ] )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( [ "A", "G", "A.G", "B.B", "A.R", "B.R", "B", "R", "A.B", "B.G" ] ), [ "R", "G", "B", "A", "A.R", "A.G", "A.B", "B.R", "B.G", "B.B" ] )
 
 		# Sort hierarchical layers
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( [ "Y.X.Y", "Y.X.X", "Y", "X", "X.X", "X.Y" ] ), [ "X", "Y", "X.X", "X.Y", "Y.X.X", "Y.X.Y" ] )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( [ "Y.X.Y", "Y.X.X", "Y", "X", "X.X", "X.Y" ] ), [ "X", "Y", "X.X", "X.Y", "Y.X.X", "Y.X.Y" ] )
 
 		# Default alphabetical sort puts capital letters before lowercase ( dunno if this is good or not, but
 		# worth documenting )
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( [ "x", "X", "x.x", "X.X", "X.x", "x.X" ] ), ["X", "x", "X.X", "X.x", "x.X", "x.x"] )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( [ "x", "X", "x.x", "X.X", "X.x", "x.X" ] ), ["X", "x", "X.X", "X.x", "x.X", "x.x"] )
 
 		# Test that our natural sort handles numbers
 		self.assertEqual(
-			GafferImage.ImageAlgo.sortChannelNames( [ "12", "4", "1", "14", "2", "3", "13", "99" ] ),
+			GafferImage.ImageAlgo.sortedChannelNames( [ "12", "4", "1", "14", "2", "3", "13", "99" ] ),
 			[ "1", "2", "3", "4", "12", "13", "14", "99" ]
 		)
 
 		self.assertEqual(
-			GafferImage.ImageAlgo.sortChannelNames( [ str( 999 - i ) for i in range( 1000 ) ] ),
+			GafferImage.ImageAlgo.sortedChannelNames( [ str( 999 - i ) for i in range( 1000 ) ] ),
 			[ str( i ) for i in range( 1000 ) ]
 		)
 
 		# And numbers with varying padding
 		self.assertEqual(
-			GafferImage.ImageAlgo.sortChannelNames( [ "03", "00012", "4", "01", "0000014", "2", "00003", "13", "3", "99" ] ),
+			GafferImage.ImageAlgo.sortedChannelNames( [ "03", "00012", "4", "01", "0000014", "2", "00003", "13", "3", "99" ] ),
 			[ "01", "2", "3", "03", "00003", "4", "00012", "13", "0000014", "99" ]
 		)
 
 		# Test large numbers ( larger than the size of an integer )
 		largeNums = [ str( ( i % 9 ) + 1 ) + "0" * i for i in range( 30 ) ]
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( largeNums ), largeNums )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( largeNums ), largeNums )
 
 		# Test something like how we see this being used in production
 		realistic = [ "channel0", "channel1", "channel2", "channel2alt", "channel2alt2", "channel2alt13", "channel04", "channel13" ]
-		self.assertEqual( GafferImage.ImageAlgo.sortChannelNames( realistic ), realistic )
+		self.assertEqual( GafferImage.ImageAlgo.sortedChannelNames( realistic ), realistic )
 
 		# And test every length 4 possibility for alternating letters and digits
 		permutations = [ "".join( i ) for j in [1,2,3,4 ] for i in itertools.product( *["012acC"] * j ) ]
-		permutationsGaffer = GafferImage.ImageAlgo.sortChannelNames( permutations )
+		permutationsGaffer = GafferImage.ImageAlgo.sortedChannelNames( permutations )
 
 
 		# The most debatable order here is between "00C0" and "0C00", which conceptually mean the same thing, and